apps/hf-api/requirements.txt
torch
torchvision
pillow
transformers>=4.39.1
pillow-heif
numpy<2
accelerate
huggingface-hub>0.25
